Move clade helper functions to Clade.{h,cpp}
These could be used in other places.
